+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.14.3](https://github.com/naimo84/node-red-contrib-ical-events/compare/v0.14.2...v0.14.3) (2020-04-07)
+
+
+### Bug Fixes
+
+* issue [#44](https://github.com/naimo84/node-red-contrib-ical-events/issues/44), synology caldav server, added scrapegoat as fallback ([4fa8df6](https://github.com/naimo84/node-red-contrib-ical-events/commit/4fa8df69984a8d4161f24fe9129061d9e760b07e))
+
+### [0.14.2](https://github.com/naimo84/node-red-contrib-ical-events/compare/v0.14.1...v0.14.2) (2020-04-06)
+
+
+### Bug Fixes
+
+* issue [#44](https://github.com/naimo84/node-red-contrib-ical-events/issues/44), synology caldav server, added scrapegoat as fallback ([972e583](https://github.com/naimo84/node-red-contrib-ical-events/commit/972e5832bc20fbefc83c9ef3eb8199e504005092))
+* issue [#45](https://github.com/naimo84/node-red-contrib-ical-events/issues/45), wrong rrule events ([cc5709f](https://github.com/naimo84/node-red-contrib-ical-events/commit/cc5709febf9c6fe0e1335c7f9a94c12b91d62e4b))
+
 ### [0.14.1](https://github.com/naimo84/node-red-contrib-ical-events/compare/v0.14.0...v0.14.1) (2020-03-27)